refactor(java): move offset field inside relevant method body and refactor unit test files
Move the character offset field to the isLatin() method body.
Move the unit tests to relevant string utils test file.
