+++ java/fury-core/src/main/java/org/apache/fury/meta/MetaStringEncoder.java
@@ -22,7 +22,6 @@ package org.apache.fury.meta;
 import java.nio.charset.StandardCharsets;
 import java.util.HashSet;
 import org.apache.fury.collection.Collections;
-import org.apache.fury.memory.Platform;
 import org.apache.fury.meta.MetaString.Encoding;
 import org.apache.fury.util.Preconditions;
 import org.apache.fury.util.StringUtils;
@@ -58,7 +57,7 @@ public class MetaStringEncoder {
     if (input.isEmpty()) {
       return new MetaString(input, Encoding.UTF_8, specialChar1, specialChar2, new byte[0]);
     }
-    if (!StringUtils.isLatin(input.toCharArray(), Platform.CHAR_ARRAY_OFFSET)) {
+    if (!StringUtils.isLatin(input.toCharArray())) {
       return new MetaString(
           input,
           Encoding.UTF_8,
@@ -80,8 +79,7 @@ public class MetaStringEncoder {
   public MetaString encode(String input, Encoding encoding) {
     Preconditions.checkArgument(
         input.length() < Short.MAX_VALUE, "Long meta string than 32767 is not allowed");
-    if (encoding != Encoding.UTF_8
-        && !StringUtils.isLatin(input.toCharArray(), Platform.CHAR_ARRAY_OFFSET)) {
+    if (encoding != Encoding.UTF_8 && !StringUtils.isLatin(input.toCharArray())) {
       throw new IllegalArgumentException("Non-ASCII characters in meta string are not allowed");
     }
     if (input.isEmpty()) {
