+++ java/fury-core/src/main/java/org/apache/fury/serializer/Serializers.java
@@ -258,7 +258,7 @@ public class Serializers {
         buffer.writeBytes(v, 0, bytesLen);
       } else {
         char[] v = (char[]) GET_VALUE.apply(value);
-        if (StringUtils.isLatin(v, Platform.CHAR_ARRAY_OFFSET)) {
+        if (StringUtils.isLatin(v)) {
           stringSerializer.writeCharsLatin(buffer, v, value.length());
         } else {
           stringSerializer.writeCharsUTF16(buffer, v, value.length());
